Create abstract class Person
