+++ src/com/journal/Main.java
@@ -1,7 +1,11 @@
 package com.journal;
 
+
+
 public class Main {
     public static void main(String[] args) {
         System.out.println("Hello world");
+
+
     }
 }
